+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased Changes
 
+### Scopes
+
+- Removed scope `app-engine:functions:run` as it's not needed
+
 ## 0.11.0
 
 - Fixed usage percentage to no longer be printed when no budget is set