+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased Changes
 
+### Scopes
+
+- Removed scope `app-engine:functions:run` as it's not needed
 - Breaking: Changed default HTTP server host binding from `0.0.0.0` to `127.0.0.1` for improved security
 
 ## 0.11.0